Add GetFilePathInHomeDirectory to userutils
